Factors explain support out into a separate module which is now mixed in to postgres and mysql
