+++ lib/arjdbc/jdbc.rb
@@ -1,2 +1,4 @@
 require 'arjdbc/jdbc/adapter'
+require 'arjdbc/jdbc/explain_support'
 require 'arjdbc/jdbc/discover'
+
